// apps/autoops-web/generated/api.ts
export interface ProviderFieldSchema {
  name?: string;
  label?: string;
  type?: string;
  required?: boolean;
  /** @nullable */
  description?: string | null;
}

export interface ProviderSchema {
  id?: string;
  name?: string;
  type?: string;
  category?: string;
  /** @nullable */
  description?: string | null;
  isActive?: boolean;
  createdAt?: string;
  updatedAt?: string;
}
// ...
export interface ListDnsProvidersItemSchema {
  code?: string;
  name?: string;
  credentialFields?: ProviderFieldSchema[];
}

export interface ListDnsProvidersResponse {
  code?: number;
  success?: boolean;
  msg?: string;
  data?: ListDnsProvidersItemSchema[];
}

export interface ListDeployTargetsItemSchema {
  type?: string;
  name?: string;
  description?: string;
}

export interface ListDeployTargetsResponse {
  code?: number;
  success?: boolean;
  msg?: string;
  data?: ListDeployTargetsItemSchema[];
}

export interface ListCertificatesDataItemDnsProviderSchema {
  id?: string;
  name?: string;
  type?: string;
}

/**
 * @nullable
 */
export type ListCertificatesDataItemDeployProviderSchema = {
  id?: string;
  name?: string;
  type?: string;
} | null;

export interface ListCertificatesDataItemCountSchema {
  logs?: number;
}

/**
 * @nullable
 */
export type ListCertificatesDataItemSchemaDeployConfig = { [key: string]: unknown } | null;

export interface ListCertificatesDataItemSchema {
  id?: string;
  name?: string;
  domain?: string;
  wildcard?: boolean;
  sanDomains?: string[];
  dnsProviderId?: string;
  deployTarget?: string;
  /** @nullable */
  deployProviderId?: string | null;
  /** @nullable */
  deployConfig?: ListCertificatesDataItemSchemaDeployConfig;
  autoRenew?: boolean;
  status?: string;
  /** @nullable */
  certPem?: string | null;
  /** @nullable */
  keyPem?: string | null;
  /** @nullable */
  chainPem?: string | null;
  /** @nullable */
  issuedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt?: string;
  updatedAt?: string;
  dnsProvider?: ListCertificatesDataItemDnsProviderSchema;
  deployProvider?: ListCertificatesDataItemDeployProviderSchema | null;
  _count?: ListCertificatesDataItemCountSchema;
}
// ...
export type GetCertificateDataLogItemSchemaDetails = { [key: string]: unknown } | null;

export interface GetCertificateDataLogItemSchema {
  id?: string;
  certificateId?: string;
  level?: string;
  message?: string;
  /** @nullable */
  details?: GetCertificateDataLogItemSchemaDetails;
  createdAt?: string;
}
// ...
export type CreateCertificateBodyDeployConfig = { [key: string]: unknown };

export type CreateCertificateBody = {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  domain: string;
  wildcard?: boolean;
  sanDomains?: string[];
  dnsProviderId: string;
  deployTarget?: string;
  deployProviderId?: string;
  deployConfig?: CreateCertificateBodyDeployConfig;
  autoRenew?: boolean;
};